Remove x, y properties in favor of rect
Rect/image are referenced frequently enough that using properties
to access them are a performance hit.  This patch removes access
to sprites with .x .y in favor of direct .rect.x .rect.y syntax
except for MousePointer which adds x & y properties since mouse
access is less frequent, and it is convenient to not think
of the mouse as a sprite.